[rpi] disable bootdelay
Having uboot boot delay can cause issues on devices whose UARTs or
serial consoles when not utilized float or otherwise are capable of
emitting phantom characters (that would trip the press any key to break
into a uboot shell) and cause the system to not boot to Linux. This was
observed on RPI5 where it was first noticed that the debug cable being
plugged into another host or not impacted boot success.